# services/payment/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.PaymentMethodEntity:
    properties:
      account_name:
        example: PT YBB Platform
        type: string
      account_number:
        example: "1234567890"
        type: string
      admin_instructions:
        example: Cek mutasi bank tanggal sekian
        type: string
      bank_name:
        description: '--- Manual Payment Fields ---'
        example: BCA
        type: string
      code:
        description: e.g., "bank_bca", "midtrans", "gopay"
        type: string
      created_at:
        description: '--- Timestamps ---'
        example: "2025-12-01T10:00:00Z"
        type: string
      description:
        type: string
      displayName:
        description: User-facing name
        type: string
      gateway_name:
        description: '--- Automatic Payment Fields ---'
        example: ""
        type: string
      gateway_type:
        example: ""
        type: string
      icon:
        description: Icon URL or identifier
        type: string
      id:
        type: string
      instructions:
        example: Silakan transfer sesuai nominal unik
        type: string
      isActive:
        type: boolean
      name:
        description: e.g., "Bank BCA", "Midtrans", "GoPay"
        type: string
      requires_proof:
        example: true
        type: boolean
      sort_order:
        example: 1
        type: integer
      type:
        $ref: '#/definitions/entities.PaymentMethodType'
      updated_at:
        example: "2025-12-01T10:00:00Z"
        type: string
    type: object
  entities.PaymentMethodType:
    enum:
    - automatic
    - manual
    type: string
    x-enum-comments:
      MethodTypeAutomatic: Via payment gateway
      MethodTypeManual: Manual verification
    x-enum-descriptions:
    - Via payment gateway
    - Manual verification
    x-enum-varnames:
    - MethodTypeAutomatic
    - MethodTypeManual
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ini adalah dokumentasi API untuk layanan pembayaran YBB Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: YBB Payment Service API
  version: "1.0"
paths:
  /payment-methods:
    get:
      description: Mendapatkan daftar semua metode pembayaran yang tersedia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/entities.PaymentMethodEntity'
              type: array
            type: object
      summary: Lihat Semua Metode Pembayaran
      tags:
      - Payment Methods
    post:
      consumes:
      - application/json
      description: Menambahkan bank atau metode pembayaran baru ke database
      parameters:
      - description: Data Payment Method
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.PaymentMethodEntity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/entities.PaymentMethodEntity'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Tambah Metode Pembayaran Baru
      tags:
      - Payment Methods
  /payment-methods/{id}:
    delete:
      description: Menghapus (Soft Delete) metode pembayaran berdasarkan ID
      parameters:
      - description: Payment Method ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Hapus Metode Pembayaran
      tags:
      - Payment Methods
    get:
      description: Mendapatkan detail metode pembayaran berdasarkan ID
      parameters:
      - description: Payment Method ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/entities.PaymentMethodEntity'
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Detail Metode Pembayaran
      tags:
      - Payment Methods
    put:
      consumes:
      - application/json
      description: Mengubah data bank atau metode pembayaran berdasarkan ID
      parameters:
      - description: Payment Method ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.PaymentMethodEntity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/entities.PaymentMethodEntity'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update Metode Pembayaran
      tags:
      - Payment Methods
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"